refactor: extract shared orchestrate parse helpers to dedicated module Deduplicate parseIntFlag, CommonOrchestrateFlags, and parseCommonOrchestrateFlag from orchestrate.ts and orchestrate-interactive.ts into orchestrate-parse-helpers.ts. Co-Authored-By: Claude <noreply@anthropic.com>
