+++ src/cli/commands/orchestrate.ts
@@ -6,7 +6,6 @@
 
 import { bootstrap } from '../../bootstrap.js';
 import type { AgentProvider } from '../../core/agents.js';
-import { AGENT_PROVIDERS, isAgentProvider } from '../../core/agents.js';
 import type { Container } from '../../core/container.js';
 import type { LoopId } from '../../core/domain.js';
 import { OrchestratorId, OrchestratorStatus } from '../../core/domain.js';
@@ -25,6 +24,7 @@ import {
   type OrchestrateInteractiveParsed,
   parseOrchestrateInteractiveArgs,
 } from './orchestrate-interactive.js';
+import { type CommonOrchestrateFlags, parseCommonOrchestrateFlag, parseIntFlag } from './orchestrate-parse-helpers.js';
 
 // ============================================================================
 // Helpers
@@ -91,13 +91,6 @@ export function waitForLoopCompletion(container: Container, loopId: LoopId): Pro
 // Arg parsing — pure functions
 // ============================================================================
 
-/** Parse and validate an integer flag value within [min, max]. */
-function parseIntFlag(name: string, value: string, min: number, max: number): Result<number, string> {
-  const val = parseInt(value, 10);
-  if (isNaN(val) || val < min || val > max) return err(`${name} must be ${min}-${max}`);
-  return ok(val);
-}
-
 interface OrchestrateCreateParsed {
   readonly kind: 'create';
   readonly goal: string;
@@ -145,69 +138,6 @@ type OrchestrateParsed =
   | OrchestrateCancelParsed
   | OrchestrateInitParsed
   | OrchestrateInteractiveParsed;
-
-/** Shared mutable state accumulated by parseCommonOrchestrateFlag. */
-interface CommonOrchestrateFlags {
-  workingDirectory: string | undefined;
-  agent: AgentProvider | undefined;
-  model: string | undefined;
-  maxDepth: number | undefined;
-  maxWorkers: number | undefined;
-  goalWords: string[];
-}
-
-/**
- * Parse a single arg from the common flag set shared between `create` and `init`.
- * Returns the new index (incremented when the flag consumed a following value),
- * or an Err if the flag is invalid. Returns `null` when the arg is not a common
- * flag — the caller should then handle it as a subcommand-specific flag or error.
- *
- * DECISION: Single-arg dispatch rather than a full loop keeps each caller in
- * control of its own iteration and unique flags (--foreground, --system-prompt, etc.).
- */
-function parseCommonOrchestrateFlag(
-  arg: string,
-  args: readonly string[],
-  i: number,
-  state: CommonOrchestrateFlags,
-): Result<number, string> | null {
-  if (arg === '--working-directory' || arg === '-w') {
-    const next = args[i + 1];
-    if (!next || next.startsWith('-')) return err('--working-directory requires a path');
-    state.workingDirectory = next;
-    return ok(i + 1);
-  }
-  if (arg === '--agent' || arg === '-a') {
-    const next = args[i + 1];
-    if (!next || next.startsWith('-')) return err(`--agent requires a name (${AGENT_PROVIDERS.join(', ')})`);
-    if (!isAgentProvider(next)) return err(`Unknown agent: "${next}". Available: ${AGENT_PROVIDERS.join(', ')}`);
-    state.agent = next;
-    return ok(i + 1);
-  }
-  if (arg === '--model' || arg === '-m') {
-    const next = args[i + 1];
-    if (!next || next.startsWith('-')) return err('--model requires a model name (e.g. claude-opus-4-5)');
-    state.model = next;
-    return ok(i + 1);
-  }
-  if (arg === '--max-depth') {
-    const parsed = parseIntFlag('--max-depth', args[i + 1], 1, 10);
-    if (!parsed.ok) return parsed;
-    state.maxDepth = parsed.value;
-    return ok(i + 1);
-  }
-  if (arg === '--max-workers') {
-    const parsed = parseIntFlag('--max-workers', args[i + 1], 1, 20);
-    if (!parsed.ok) return parsed;
-    state.maxWorkers = parsed.value;
-    return ok(i + 1);
-  }
-  if (!arg.startsWith('-')) {
-    state.goalWords.push(arg);
-    return ok(i);
-  }
-  return null; // not a common flag — caller decides
-}
 
 export function parseOrchestrateCreateArgs(args: readonly string[]): Result<OrchestrateCreateParsed, string> {
   const state: CommonOrchestrateFlags = {
